requirements.txt
absl-py==1.4.0
accelerate==0.20.3
aiohttp==3.8.5
aiosignal==1.3.1
albumentations==0.4.3
altair==5.0.1
antlr4-python3-runtime==4.9.3
anyio==3.7.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
bitarray==2.7.5
black==23.10.1
bleach==6.0.0
blinker==1.6.2
blis==0.7.9
braceexpand==0.1.7
brotlipy==0.7.0
cachetools==5.3.1
catalogue==2.0.8
cchardet==2.1.7
certifi==2022.12.7
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
clip 
cloudpickle==2.2.1
cmake==3.25.0
colorama==0.4.6
comm==0.1.3
confection==0.0.4
constantly==23.10.4
contexttimer==0.3.3
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
Cython==0.29.35
datasets==2.12.0
debugpy==1.6.7
decord==0.6.0
defusedxml==0.7.1
diffusers
dill==0.3.6
distlib==0.3.6
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.6.1
exceptiongroup==1.1.1
fairscale==0.4.4
fastjsonschema==2.17.1
packaging==23.2
torch-fidelity==0.3.0
torchmetrics==0.5.0
fonttools==4.43.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.10.0
ftfy==6.1.1
future==0.18.3
fvcore==0.1.5.post20221221
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.33
google-auth==2.19.1
google-auth-oauthlib==1.0.0
grpcio==1.54.2
hkdf==0.0.3
huggingface-hub==0.16.4
humanize==4.9.0
Hydra==2.5
hydra-core==1.3.2
hydra-submitit-launcher==1.2.0
hyperlink==21.0.0
identify==2.5.24
idna==3.4
imageio==2.9.0
imageio-ffmpeg==0.4.2
imgaug==0.2.6
importlib-metadata==6.6.0
importlib-resources==5.12.0
incremental==22.10.0
iopath==0.1.9
ipykernel==6.24.0
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
Jinja2==3.1.3
joblib==1.3.1
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kaggle==1.5.13
kiwisolver==1.4.4
kornia==0.5.8
langcodes==3.3.0
lazy_loader==0.2
lightning-utilities==0.9.0
lit==15.0.7
lvis==0.5.3
lxml==4.9.2
magic-wormhole==0.13.0
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
mistune==2.0.5
mkl-fft==1.3.6
mkl-service==2.4.0
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.5.0
nbformat==5.9.0
ninja==1.11.1
nltk==3.8.1
nodeenv==1.8.0
notebook==6.5.4
notebook_shim==0.2.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.2.140
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
open-clip-torch==2.20.0
opencv-python==4.7.0.72
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
overrides==7.3.1
pandas==2.0.2
pandocfilters==1.5.0
pathspec==0.11.2
pathtools==0.1.2
pathy==0.10.2
Pillow==9.0.1
pipreqs==0.4.13
pkgutil_resolve_name==1.3.10
plotly==5.15.0
portalocker==2.7.0
pre-commit==3.3.3
preshed==3.0.8
prometheus-client==0.17.0
protobuf==3.20.3
pudb==2019.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.6
pydantic==1.10.9
pydeck==0.8.1b0
pyDeprecate==0.3.1
Pympler==1.0.1
PyNaCl==1.5.0
pyparsing==3.0.9
pyre-extensions==0.0.29
pyrsistent==0.19.3
python-json-logger==2.0.7
python-magic==0.4.27
python-slugify==8.0.1
pytorch-fid==0.3.0
pytorch-lightning==1.4.2
pytorch-triton
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
regex==2023.10.3
requests-oauthlib==1.3.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.1
rsa==4.9
sacrebleu==2.3.1
safetensors==0.3.3
salesforce-lavis
scikit-image==0.20.0
scipy==1.9.1
seaborn==0.12.2
Send2Trash==1.8.2
sentencepiece==0.1.99
sentry-sdk==1.25.1
service-identity==23.1.0
setproctitle==1.3.2
shapely==2.0.1
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spake2==0.8
srsly==2.4.6
streamlit==1.23.1
submitit==1.4.5
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.0
termcolor==2.3.0
terminado==0.17.1
test-tube==0.7.5
text-unidecode==1.3
thinc==8.1.10
tifffile==2023.4.12
timm==0.4.12
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
transformers==4.27.0
triton==2.0.0
Twisted==23.10.0
txaio==23.1.1
txtorcon==23.11.0
typer==0.7.0
typing-inspect==0.9.0
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==4.3
ultralytics==8.0.119
uri-template==1.2.0
urwid==2.1.2
validators==0.20.0
virtualenv==20.23.1
wandb==0.15.4
wasabi==1.1.2
watchdog==3.0.0
webcolors==1.13
webdataset==0.2.48
webencodings==0.5.1
websocket-client==1.5.3
Werkzeug==2.3.8
widgetsnbextension==4.0.7
xformers==0.0.20
xxhash==3.2.0
yacs==0.1.8
yarg==0.1.9
yarl==1.9.2
zipp==3.15.0
zope.interface==6.1
